pyproject.toml
[tool.poetry]
name = "saltybet-asyncio"
version = "0.1.0"
description = ""
authors = ["Nick Pancakes <nick@pncks.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-socketio = {extras = ["asyncio_client"], version = "^4.6.0"}
selectolax = "^0.2.10"
pendulum = "^2.1.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pylint = "^2.5.3"
mypy = "^0.800"

[build-system]
requires = ["setuptools", "poetry>=0.12"]
build-backend = "poetry.masonry.api"
